+++ application/views/equipment.php
@@ -1,8 +1,2 @@
-<div class="row">
-	<div class="span2">
-                <img src="/pix/{image}.jpg" title="{code}"/><br/>
-                code {code}
-                name {name}
-                attributes {attributes}
-	</div>
-</div>
+<!-- author: Michaela -->
+{thetable}